Set document resolution (partial implementation)

+++ DigitNow.Microservice.DocumentManagement/DigitNow.Domain.DocumentManagement/Business/IncomingDocuments/Commands/Update/UpdateIncomingDocumentCommand.cs
@@ -24,5 +24,5 @@ public class UpdateIncomingDocumentCommand : ICommand<ResultObject>
     public double ResolutionPeriod { get; set; }
     public bool? IsUrgent { get; set; }
     public bool? IsGDPRAgreed { get; set; }
-    public List<int> ConnectedDocumentIds { get; set; }
+    public List<long> ConnectedDocumentIds { get; set; }
 }
